Refactor pre-validations to remove validator factory and abstract validator
