--- migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/validator/validators/ValidatorFactory.java
@@ -1,19 +1,0 @@
-package org.wso2.carbon.apimgt.migration.validator.validators;
-
-
-import org.wso2.carbon.apimgt.migration.validator.utils.Utils;
-
-public class ValidatorFactory {
-    private final Utils utils;
-
-    public ValidatorFactory(Utils utils) {
-        this.utils = utils;
-    }
-
-    public Validator getVersionValidator(String migrateFromVersion) {
-        if ("4.0.0".equals(migrateFromVersion)) {
-            return new V400Validator(utils);
-        }
-        return null;
-    }
-}
